+++ zarrs_filesystem/tests/filesystem.rs
@@ -1,0 +1,56 @@
+#![allow(missing_docs)]
+
+use zarrs_filesystem::FilesystemStore;
+
+use std::error::Error;
+
+#[cfg(target_os = "linux")]
+fn try_open_direct_io(path: &str) -> std::io::Result<std::fs::File> {
+    use libc::{open, O_DIRECT, O_RDONLY};
+    use std::os::fd::FromRawFd;
+
+    let c_path = std::ffi::CString::new(path).unwrap();
+    unsafe {
+        let fd = open(c_path.as_ptr(), O_RDONLY | O_DIRECT);
+        if fd < 0 {
+            Err(std::io::Error::last_os_error())
+        } else {
+            Ok(std::fs::File::from_raw_fd(fd))
+        }
+    }
+}
+
+#[test]
+#[cfg_attr(miri, ignore)]
+fn filesystem() -> Result<(), Box<dyn Error>> {
+    let path = tempfile::TempDir::new()?;
+    let store = FilesystemStore::new(path.path())?.sorted();
+    // let store = FilesystemStore::new("tests/data/store")?.sorted();
+    zarrs_storage::store_test::store_write(&store)?;
+    zarrs_storage::store_test::store_read(&store)?;
+    zarrs_storage::store_test::store_list(&store)?;
+    Ok(())
+}
+
+#[cfg(target_os = "linux")]
+#[test]
+// #[cfg_attr(miri, ignore)]
+fn direct_io() -> Result<(), Box<dyn Error>> {
+    use zarrs_filesystem::FilesystemStoreOptions;
+
+    let tmpfile = tempfile::NamedTempFile::new()?;
+    if try_open_direct_io(tmpfile.path().to_str().unwrap()).is_err() {
+        // Skip this test if direct I/O is not supported
+        return Ok(());
+    }
+
+    let path = tempfile::TempDir::new()?;
+    let mut opts = FilesystemStoreOptions::default();
+    opts.direct_io(true);
+
+    let store = FilesystemStore::new_with_options(path.path(), opts)?.sorted();
+    zarrs_storage::store_test::store_write(&store)?;
+    zarrs_storage::store_test::store_read(&store)?;
+    zarrs_storage::store_test::store_list(&store)?;
+    Ok(())
+}